Followers now imprint on Leaders
When a follower spawns, it will now attempt to find the closest leader and follow it around. This check runs at the end of every timer round if a leader is not found.

This may result in ducklings following around the player if they get too close when one is born